+++ Features/Ui/ExaniteEngineMyraPlatform.cs
@@ -1,14 +1,5 @@
-using System;
 using System.Collections.Generic;
 using System.Drawing;
-using System.Numerics;
-using Diligent;
-using Exanite.Engine.Rendering;
-using Exanite.GravitationalTetris.Features.Rendering;
-using Exanite.GravitationalTetris.Features.Sprites.Systems;
-using FontStashSharp;
-using FontStashSharp.Interfaces;
-using Myra.Graphics2D;
 using Myra.Graphics2D.UI;
 using Myra.Platform;
 using Window = Exanite.Engine.Windowing.Window;
@@ -56,110 +47,5 @@ public class ExaniteEngineMyraPlatform : IMyraPlatform
     public TouchCollection GetTouchState()
     {
         return touchState;
-    }
-}
-
-public class ExaniteEngineMyraRenderer : IMyraRenderer
-{
-    private readonly SpriteBatchSystem spriteBatchSystem;
-
-    public ExaniteEngineMyraRenderer(ITexture2DManager textureManager, SpriteBatchSystem spriteBatchSystem)
-    {
-        this.spriteBatchSystem = spriteBatchSystem;
-
-        TextureManager = textureManager;
-    }
-
-    public ITexture2DManager TextureManager { get; }
-
-    public RendererType RendererType => RendererType.Sprite;
-    public Rectangle Scissor { get; set; }
-
-    public void Begin(TextureFiltering textureFiltering)
-    {
-        // Todo No batching/instancing for now
-    }
-
-    public void End()
-    {
-        // Todo No batching/instancing for now
-    }
-
-    public void DrawSprite(object texture, Vector2 pos, Rectangle? src, FSColor color, float rotation, Vector2 scale, float depth)
-    {
-        var typedTexture = (Texture2D)texture;
-
-        spriteBatchSystem.DrawSprite(typedTexture, new SpriteUniformData
-        {
-            World = Matrix4x4.Identity,
-            View = Matrix4x4.Identity,
-            Projection = Matrix4x4.Identity,
-            Color = Vector4.One,
-        });
-    }
-
-    public void DrawQuad(object texture, ref VertexPositionColorTexture topLeft, ref VertexPositionColorTexture topRight, ref VertexPositionColorTexture bottomLeft, ref VertexPositionColorTexture bottomRight)
-    {
-        throw new NotSupportedException();
-    }
-}
-
-public class ExaniteEngineFontTextureManager : ITexture2DManager, IDisposable
-{
-    private int nextId = 0;
-    private readonly List<Texture2D> textures = new();
-
-    private readonly RendererContext rendererContext;
-
-    public ExaniteEngineFontTextureManager(RendererContext rendererContext)
-    {
-        this.rendererContext = rendererContext;
-    }
-
-    public object CreateTexture(int width, int height)
-    {
-        var texture = new Texture2D($"FontStashSharp Texture #{nextId}", width, height, rendererContext, Usage.Dynamic);
-        textures.Add(texture);
-
-        nextId++;
-
-        return texture;
-    }
-
-    public Point GetTextureSize(object texture)
-    {
-        var typedTexture = (Texture2D)texture;
-
-        return new Point(typedTexture.Width, typedTexture.Height);
-    }
-
-    public unsafe void SetTextureData(object texture, Rectangle bounds, byte[] data)
-    {
-        var typedTexture = (Texture2D)texture;
-
-        var resource = rendererContext.DeviceContext.MapTextureSubresource(typedTexture.Texture, 0, 0, MapType.Write, MapFlags.Discard, new Box
-        {
-            MinX = (uint)bounds.X,
-            MinY = (uint)bounds.Y,
-            MaxX = (uint)(bounds.X + bounds.Width),
-            MaxY = (uint)(bounds.Y + bounds.Height),
-        });
-        {
-            var resourceData = new Span<byte>((void*)resource.Data, bounds.Width * bounds.Height);
-            // Todo Going to naively copy for now - Likely need to change RGBA order
-            for (var i = 0; i < data.Length; i++)
-            {
-                resourceData[i] = data[i];
-            }
-        }
-        rendererContext.DeviceContext.UnmapTextureSubresource(typedTexture.Texture, 0, 0);
-    }
-
-    public void Dispose()
-    {
-        foreach (var texture in textures)
-        {
-            texture.Dispose();
-        }
     }
 }
